src/packets/requirements.txt
black
jinja2==3.0.3
pyyaml==6.0.0
rich
